feat(pipeline): add REST GET scrape dispatch
Add a GET verb to the api-direct-scrape hard model so browser banks
whose whole API contract is GET (Discount/Titan family) can dispatch
accounts/balance/transactions without a request body.

- IApiDirectScrapeShape: new ScrapeHttpMethod ('GET' | 'POST') type +
  optional method on all three steps; defaults to POST so every
  existing bank stays inert.
- ApiDirectScrapeDispatch: dispatchStep is now a 3-way router
  (GraphQL -> apiQuery, GET -> apiGet, POST -> apiPost).
- Split ApiDirectScrapeSteps into a new lower-level
  ApiDirectScrapeDispatchArgs module (owns the driver/account context
  types + dispatch-arg factories) so Steps stays under the 150-line
  cap after threading method. Import graph stays one-directional
  (Actions -> Steps -> DispatchArgs); no cycle.
- Docs + docs-coverage allowlist updated for the moved internals.

No behaviour change for existing banks; framework prep for the
Discount pilot.
