+++ src/Scrapers/Pipeline/Phases/ApiDirectScrape/ApiDirectScrapeActions.ts
@@ -16,13 +16,12 @@ import { isSome, some } from '../../Types/Option.js';
 import type { IActionContext, IScrapeState } from '../../Types/PipelineContext.js';
 import type { Procedure } from '../../Types/Procedure.js';
 import { isOk, succeed } from '../../Types/Procedure.js';
+import type { IAcctCtx, IDriverCtx } from './ApiDirectScrapeDispatchArgs.js';
 import {
   buildPageFetcher,
   buildStop,
   fetchAccounts,
   fetchBalance,
-  type IAcctCtx,
-  type IDriverCtx,
 } from './ApiDirectScrapeSteps.js';
 import type { ApiDirectScrapeResult } from './ApiDirectScrapeTypes.js';
 import type { IApiDirectScrapeShape } from './IApiDirectScrapeShape.js';
